pyproject.toml
[tool.poetry]
name = "Kyoyobot"
version = "0.1.0"
description = "🐛"
authors = ["96 LB <96lb@96lb.ml>", "A bunch of other people probably"]

[tool.poetry.dependencies]
python = ">=3.8.0"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
Flask = "^2.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
